split syntax_checker
